// frontend/src/api-generated/models/keHoachBaoTriDinhKyRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietKeHoachBaoTriRequest } from './chiTietKeHoachBaoTriRequest';

export interface KeHoachBaoTriDinhKyRequest {
  tenKeHoach: string;
  chuKyLap: string;
  thoiGianBatDauKeHoach: string;
  thoiGianKetThucKeHoach: string;
  chiPhiDuKien?: number;
  noiDungBaoTri?: string;
  danhSachChiTiet?: ChiTietKeHoachBaoTriRequest[];
}